tests/wpt/web-platform-tests/offscreen-canvas/pixel-manipulation/2d.imageData.get.nonfinite.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.imageData.get.nonfinite</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>

<h1>2d.imageData.get.nonfinite</h1>
<p class="desc">getImageData() throws TypeError if arguments are not finite</p>


<script>
var t = async_test("getImageData() throws TypeError if arguments are not finite");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

assert_throws(new TypeError(), function() { ctx.getImageData(Infinity, 10, 10, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(-Infinity, 10, 10, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(NaN, 10, 10, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, Infinity, 10, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, -Infinity, 10, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, NaN, 10, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, 10, Infinity, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, 10, -Infinity, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, 10, NaN, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, 10, 10, Infinity); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, 10, 10, -Infinity); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, 10, 10, NaN); });
assert_throws(new TypeError(), function() { ctx.getImageData(Infinity, Infinity, 10, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(Infinity, Infinity, Infinity, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(Infinity, Infinity, Infinity, Infinity); });
assert_throws(new TypeError(), function() { ctx.getImageData(Infinity, Infinity, 10, Infinity); });
assert_throws(new TypeError(), function() { ctx.getImageData(Infinity, 10, Infinity, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(Infinity, 10, Infinity, Infinity); });
assert_throws(new TypeError(), function() { ctx.getImageData(Infinity, 10, 10, Infinity); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, Infinity, Infinity, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, Infinity, Infinity, Infinity); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, Infinity, 10, Infinity); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, 10, Infinity, Infinity); });
var posinfobj = { valueOf: function() { return Infinity; } },
    neginfobj = { valueOf: function() { return -Infinity; } },
    nanobj = { valueOf: function() { return -Infinity; } };
assert_throws(new TypeError(), function() { ctx.getImageData(posinfobj, 10, 10, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(neginfobj, 10, 10, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(nanobj, 10, 10, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, posinfobj, 10, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, neginfobj, 10, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, nanobj, 10, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, 10, posinfobj, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, 10, neginfobj, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, 10, nanobj, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, 10, 10, posinfobj); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, 10, 10, neginfobj); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, 10, 10, nanobj); });
assert_throws(new TypeError(), function() { ctx.getImageData(posinfobj, posinfobj, 10, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(posinfobj, posinfobj, posinfobj, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(posinfobj, posinfobj, posinfobj, posinfobj); });
assert_throws(new TypeError(), function() { ctx.getImageData(posinfobj, posinfobj, 10, posinfobj); });
assert_throws(new TypeError(), function() { ctx.getImageData(posinfobj, 10, posinfobj, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(posinfobj, 10, posinfobj, posinfobj); });
assert_throws(new TypeError(), function() { ctx.getImageData(posinfobj, 10, 10, posinfobj); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, posinfobj, posinfobj, 10); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, posinfobj, posinfobj, posinfobj); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, posinfobj, 10, posinfobj); });
assert_throws(new TypeError(), function() { ctx.getImageData(10, 10, posinfobj, posinfobj); });
t.done();

});
</script>
